Update square property keywords and add game models
Changed the property keywords in the squares.txt file to be more Pythonic (from camelCase to snake_case). This brings more consistency to the rest of the codebase, following PEP-8 styles.

Also introduced new components of the game - 'game.py', 'player.py', 'square.py' and 'main.py', setting up the basic infrastructure for a single round of gameplay. Added these classes to create the game logic for the Monopoly game. Each class represents a different component of the game, such as a player, a square on the game board, and the game itself. The 'main.py' is used to start the game. The models simplify game interaction and progress.